+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added EventThread::runInLoop() to run a function in the context of the event thread.
+
 ## [1.1.0] - 2025-08-06
 
 ### Added